requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deathbycaptcha-official==4.6.1
decorator==5.1.1
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
FastAPI-SQLAlchemy==0.2.1
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
outcome==1.2.0
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rsa==4.9
selenium==4.14.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.18.0
vine==5.0.0
watchfiles==0.21.0
wcwidth
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
